+++ index.js
@@ -33,6 +33,10 @@ module.exports = (context, options = {}) => {
     ...options,
   };
 
+  let reactOpts = {
+    useSpread: true,
+  };
+
   let runtimeOpts = {
     corejs: 3,
     helpers: true,
@@ -42,7 +46,7 @@ module.exports = (context, options = {}) => {
 
   let presets = [
     ['@babel/preset-env', envOpts],
-    '@babel/preset-react',
+    ['@babel/preset-react', reactOpts],
     '@babel/preset-flow',
   ];